requirements.txt
appnope==0.1.4
astropy==7.1.0
astropy-iers-data==0.2025.9.8.0.36.17
astroscrappy==1.3.0
asttokens==3.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
h5py==3.14.0
ImageIO==2.37.3
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lazy-loader==0.5
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
photutils==2.2.0
pillow==11.3.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyerfa==2.0.1.5
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
scikit-image==0.26.0
scipy==1.16.1
setuptools==80.9.0
simpleitk==2.5.3
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tifffile==2026.3.3
torch==2.8.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
wcwidth==0.2.14
widgetsnbextension==4.0.15
